// src/userIntf/code/odlux/apps/preproviderApp/src/models/preproviderServer.ts
 /***
  * ################################################################################################
  * #                                                                                              #
  * #   Copyright (C)  [2019]  [RideNext Software Solutions Pvt Ltd]. All rights  reserved         #
  * #                                                                                              #
  * ################################################################################################
 ****/

export type providerdata = {
   DeviceType: string;
   PNFID: string;
   IP_ADDRESS: string;
   PORT_NUMBER: string;
   USERNAME: string;
   PASSWORD: string;
   PREPROVIDER_CONF: string; 
   FILENAME: string; 
   oldPNFID:string;
   oldip:string;
   oldport:string;
   SSHKey:string;
}